# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  api.ConsoleCreateAPIKeyRequest:
    properties:
      name:
        maxLength: 100
        minLength: 1
        type: string
      scopes:
        items:
          type: string
        minItems: 1
        type: array
    required:
    - name
    - scopes
    type: object
...
  api.ErrorResponse:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/api.ResponseCode'
        description: response code
      error:
        description: error detail
        type: string
      errors:
        description: field errors
        items:
          $ref: '#/definitions/api.FieldError'
        type: array
      message:
        description: response message
        type: string
      request_id:
        description: request id
        type: string
      timestamp:
        description: timestamp (ms)
        type: integer
    type: object
...
  api.FieldError:
    properties:
      field:
        description: field name
        type: string
      message:
        description: error message
        type: string
      value:
        description: field value
        type: string
    type: object
...
  api.RefreshTokenRequest:
    properties:
      client_id:
        type: string
      refresh_token:
        type: string
    required:
    - client_id
    - refresh_token
    type: object
...
  api.ResponseCode:
    enum:
    - 200
    - 1000
    - 1001
    - 1002
    - 1003
    - 1004
    - 1005
    - 1006
    - 1007
    - 2000
    - 2001
    - 2002
    - 2003
    - 2004
    - 2005
    - 2006
    - 2007
    - 40201
    - 5000
    - 5001
    - 5002
    - 5003
    - 5004
    type: integer
    x-enum-comments:
      CodeBadRequest: invalid request parameters
      CodeBusinessError: generic business error
      CodeConflict: resource conflict
      CodeDatabaseError: database error
      CodeEncryptionError: encryption error
      CodeFaceNotMatch: face not matched
      CodeFaceVerifyFailed: face verification failed
      CodeForbidden: forbidden
      CodeIDCardNotMatch: id card information not matched
      CodeInternalError: internal server error
      CodeInvalidParameter: invalid parameter
      CodeKYCFailed: kyc verification failed
      CodeLivenessFailed: liveness detection failed
      CodeMethodNotAllowed: method not allowed
      CodeMissingParameter: missing parameter
      CodeNotFound: resource not found
      CodeOCRFailed: ocr recognition failed
      CodePaymentRequired: quota exceeded, please upgrade your plan
      CodeServiceUnavailable: service unavailable
      CodeThirdPartyError: third party service error
      CodeTooManyRequests: too many requests
      CodeUnauthorized: unauthorized
    x-enum-descriptions:
    - ""
    - invalid request parameters
    - unauthorized
    - forbidden
    - resource not found
    - method not allowed
    - too many requests
    - invalid parameter
    - missing parameter
    - generic business error
    - ocr recognition failed
    - face verification failed
    - liveness detection failed
    - kyc verification failed
    - id card information not matched
    - face not matched
    - resource conflict
    - quota exceeded, please upgrade your plan
    - internal server error
    - database error
    - third party service error
    - service unavailable
    - encryption error
    x-enum-varnames:
    - CodeSuccess
    - CodeBadRequest
    - CodeUnauthorized
    - CodeForbidden
    - CodeNotFound
    - CodeMethodNotAllowed
    - CodeTooManyRequests
    - CodeInvalidParameter
    - CodeMissingParameter
    - CodeBusinessError
    - CodeOCRFailed
    - CodeFaceVerifyFailed
    - CodeLivenessFailed
    - CodeKYCFailed
    - CodeIDCardNotMatch
    - CodeFaceNotMatch
    - CodeConflict
    - CodePaymentRequired
    - CodeInternalError
    - CodeDatabaseError
    - CodeThirdPartyError
    - CodeServiceUnavailable
    - CodeEncryptionError
...
  api.SuccessResponse:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/api.ResponseCode'
        description: response code
      data:
        description: response data
      message:
        description: response message
        type: string
      request_id:
        description: request id
        type: string
      timestamp:
        description: timestamp (ms)
        type: integer
    type: object
  api.TokenRequest:
    properties:
      client_id:
        type: string
      client_secret:
        type: string
      grant_type:
        type: string
      scope:
        type: string
    required:
    - client_id
    - client_secret
    - grant_type
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 企业级KYC认证服务
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: KYC Service API
  version: "1.0"
paths:
...
  /api/v1/keys:
    get:
      consumes:
      - application/json
      description: 获取当前用户的所有API密钥
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: 获取用户API密钥列表
      tags:
      - Credentials
    post:
      consumes:
      - application/json
      description: 创建新的API密钥
      parameters:
      - description: 创建API密钥请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.ConsoleCreateAPIKeyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: 创建API密钥
      tags:
      - Console
  /api/v1/keys/{id}:
    delete:
      consumes:
      - application/json
      description: 撤销指定的API密钥
      parameters:
      - description: API密钥ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: 撤销API密钥
      tags:
      - Console
...
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Use the refresh token to obtain a new access token
      parameters:
      - description: Refresh token request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SuccessResponse'
      summary: Refresh Access Token
      tags:
      - Auth
      - Public
  /auth/token:
    post:
      consumes:
      - application/json
      description: Use Client credentials to obtain access tokens
      parameters:
      - description: Token request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.TokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SuccessResponse'
      summary: Obtain access token
      tags:
      - Auth
      - Public
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"